requirements.txt
asgiref==3.11.0
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
sqlparse==0.5.4
tzdata==2025.2
pillow==12.0.0
